+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Change log
 
-## v1.7.0
+## v1.8.0 (12/20/2015)
+
+- Add `serve` and `rehash` subcommands to the CLI interface.
+- End users should now use the new subcommands to specify the mode. `serve` will be the default.
+
+
+## v1.7.0 (12/19/2015)
 
 - Add `--help` and `--version` commands
 - When no options are provided display the help menu.